Issue #3316405 by joseph.olstad: In the DateTimeBlock class $time can sometimes not be an integer and core gets offended

+++ src/Plugin/Block/DateModifiedBlock.php
@@ -175,7 +175,7 @@ class DateModifiedBlock extends BlockBase implements ContainerFactoryPluginInter
     else {
       $formatted_date = $this->dateFormatStorage->load($format)->getPattern();
     }
-    $time = (int) $time;
+    $time = (float) $time;
     $date = $this->dateFormatter->format($time, 'custom', $formatted_date);
 
     $build = [];
